Exception handling - delete

+++ src/main/java/com/rodrigopepato/course/services/UserService.java
@@ -2,8 +2,11 @@ package com.rodrigopepato.course.services;
 
 import com.rodrigopepato.course.entities.User;
 import com.rodrigopepato.course.repositories.UserRepository;
+import com.rodrigopepato.course.services.exceptions.DatabaseException;
 import com.rodrigopepato.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +32,15 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            repository.deleteById(id);
+            repository.flush();
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
